placment: add gui skeleton
We're going to need to figure out why the placer is misbehaving, this commit
adds the libraries required to get some graphical output in the placer and sets
up the command.

The test scripts are probably broken now because the command format changed but
I don't much care at the moment.