Revert "placement: AoS -> SoA conversion"
This reverts commit 36cf1e8af75847cff975ab7c6db01985483836d1.

It turns out the SoA and AoS representation are within noise, at least
on the serv core (~1.2k cells). Therefore we prefer the version that is
slightly more readable

From hyperfine,
AoS:
  Time (mean ± σ):      60.4 ms ±   0.9 ms    [User: 56.8 ms, System: 3.4 ms]
  Range (min … max):    58.5 ms …  61.7 ms    47 runs
SoA:
  Time (mean ± σ):      61.1 ms ±   1.5 ms    [User: 57.2 ms, System: 3.7 ms]
  Range (min … max):    58.9 ms …  69.3 ms    46 runs

It might be worth resurecting this code in the future if it turns out
that e.g. diffusion behaves better with SoA layouts