placement: initial velocity field computation
This will probably needs some (lots) of tweaking as we go, because mcpnr
works with much smaller cell counts and placement regions than is
typical for a "real" placement job. For example, we have a lot more
empty placement regions at the beginning and the behavior around the
edges is significantly different.

Perhaps we should expand the density and velocity fields to include 1
region outside the actual placement bounds to avoid some of the trickier
things there?